+++ packages/integration-karma/test/rendering/elements-are-not-recycled/x/child/child.html
@@ -1,5 +1,5 @@
 <template>
-    <template if:true={_isOpen}>
+    <template if:true={open}>
         <!-- If you remove the outer div, it works -->
         <div>
             <slot></slot>
