+++ packages/integration-karma/test/rendering/elements-are-not-recycled/x/child/child.js
@@ -1,0 +1,5 @@
+import { LightningElement, api } from 'lwc';
+
+export default class Child extends LightningElement {
+    @api open = false;
+}
