+++ packages/integration-karma/test/rendering/elements-are-not-recycled/x/container/container.js
