+++ packages/integration-karma/test/rendering/elements-are-not-recycled/x/simple/simple.html
