--- packages/integration-karma/test/rendering/slot-dont-recycle-vm/x/child/child.js
@@ -1,19 +1,0 @@
-import { LightningElement, track, api } from 'lwc';
-
-export default class Child extends LightningElement {
-    @track _open = false;
-
-    get _isOpen() {
-        return this._open;
-    }
-
-    @api
-    open() {
-        this._open = true;
-    }
-
-    @api
-    close() {
-        this._open = false;
-    }
-}
